design: constrain 'Premature burial' to one solution
Currently, 'Premature burial' has two similar solutions. The bottom
gravestone can be placed in one of two final positions, the bottom left
or below the bushes. I feel like the former solution is a bit harder to
see while the latter requires that the player demonstrate some
foresight. This change places a single bush, and eliminates the former
solution.