feat!: expand movement tile tileset to include diagonals and zeros (#184)
A new mechanic is being added: arrow blocks. The movement table will no
longer be static, rather, the rows and columns that its cells belong to
will be defined by the location of arrow blocks. This mechanic
introduces the possibility of overlapping columns, which can lead to
diagonal movements, or movements that cancel out. These new movements
need additional entries in the movement tileset. This adds those
entries, and adjusts the current code to cope with it before arrow
blocks are complete.