feat: add arrow block plugin (#185)
A new mechanic is being added: arrow blocks. The movement table will no
longer be static, rather, the rows and columns that its cells belong to
will be defined by the location of arrow blocks. This adds the arrow
block plugin, which is responsible for spawning arrow blocks, and
spawning the movement tiles in the cells determined by the arrow block
locations. This plugin is not actually used yet, all of the preparation
for this new feature will be completed incremementally, to eventually
replace the existing behavior wholesale.