fix: downgrade dependencies in wasm target to limit getrandom to 0.2
getrandom 0.3 requires a different set of feature flags to be enabled
for wasm compaitibility. Some dependencies of dependencies depend on
getrandom, and upgraded to getrandom 0.3 in a way that breaks bevy's
wasm compatibility. This leads to wasm build failures in our CI.
Limiting these sub-dependencies to versions that merely depend on
getrandom 0.2 should resolve the issue.