ci: run version CI against all merges instead of daily
Originally, I had the version CI that tags versions of the game and
triggers releases running daily. I thought this might be more efficient
than performing it on every merge. However, most daily attempts to
create new versions are no-ops that still take up github runner
resources, to the point that github disables our CI. This switches to a
release on every merge instead.

An attempt is being made here to not recursively trigger the version CI
workflow when the commit to increase the version in the Cargo.toml
is pushed. However, it's possible that this will also cause the release
CI to be skipped. We will test this live.