fix: only skip version CI in commits made by version CI
An attempt was made to avoid recursively processing commits pushed to
main in the version CI by using github's skip ci feature. However, this
also prevents the release ci from running on these commits, which is not
the behavior we want. This change implements a more custom solution that
uses and checks for a more specific message in the commits that the
version CI creates/should skip.