feat: add GravestoneMovementQueries system param (#190)
A new mechanic is being added: arrow blocks. The movement table will no
longer be static, rather, the rows and columns that its cells belong to
will be defined by the location of arrow blocks. This adds a system
param related to this feature: GravestoneMovementQueries. The method
provided on this system param will help reduce repeated code between
plugins that need to figure out what movements are associated with what
gravestones. This system param is not actually employed in the code yet,
all of the preparation for arrow blocks will be completed incrementally,
and a final change will replace the existing behavior wholesale.