fix: add missing styles.css file (#192)
The index file for browser builds was recently changed in an attempt to
get browser deployments working again. It now depends on a styles.css
file. However, I forgot to add this file when making the change. This
change adds that file, as well as make a dummy change to the Cargo.toml
to trigger all the necessary build processes on merge.