feat!: replace movement table with arrow blocks (#193)
Currently, the movement table is static. some work has been done making
progress towards an alternative: arrow blocks. Now, the rows and columns
that movement table's cells belong to will be defined by the location of
arrow blocks. To be more specific, an arrow block points in a direction
and is either a "row" or a "column" block, and the row or column that
the arrow block is sitting on now becomes the row or column of the
movement table associated with that direction. This allows for a lot
more expression in the shape of movement tables during level design,
making it easier to make levels where the player has access to very few
movement tiles, or to create interesting gaps in the table. Since arrow
blocks can be pushed around, more advanced levels will require the
player to do so, which can have interesting consequences (like what
happens when multiple column blocks sit in the same column).

This change finally replaces the existing movement table behavior with
arrow blocks. It also adjusts every existing level to have arrow blocks
instead of a movement table.