feat!: add debug level (#194)
It is useful to have a sandbox level to try out new features and as a
level template during development. This adds a "Debug" level in the 0th
position. It also updates the code so that, during normal play, the
players can't access this 0th level. Some other changes needed to be
made to make the level numbering appear 1-indexed again.