design: add level 'Abacus' (#168)
This is the first level that takes advantage of the movable arrows
meaningfully. It's pretty hard, and I worry that the solution might be a
little too hyper-specific, especially considering that it involves
creating a table with "diagonal" movement range. However, I think it's
worthy of taking to playtesting, at least.

![abacus](https://github.com/Trouv/willos-graveyard/assets/15664670/9c5a26e8-9b89-4d4b-925d-76218342050d)