design: explicitly introduce technique in 'Barrier magic' (#170)
The current iteration of 'Barrier magic' is a little under-designed. The
main technique it tests users on seems to come out of nowhere.
Furthermore, the level is a little unfocused, containing sub-problems
that have nothing to do with the main idea. These facts combined make
the level unfair, and players have expressed frustration with it.

This new iteration aims to address these problems by..
- forcing players to accidentally encounter the technique early on in
the puzzle
- having the solution be exclusively focused on this technique


![barrier-magic](https://github.com/Trouv/willos-graveyard/assets/15664670/0ce3a02e-200d-47f6-9cd8-dfe0e762cc1f)