design: replace level 'D row' with 'Turnover' (#172)
The main purpose of the level 'D row' is to introduce table editing. It
mostly accomplished this by significantly restricting the player's
access to the movement table in a pre-arrow-block way. The level was a
bit underdesigned and not difficult enough to feel like a real puzzle.

This PR adds a replacement level which takes advantage of movable arrow
blocks. Instead of only one row of the table being accessible, only one
tile is. With movable arrow blocks, this single tile can be
geographically separated from the others, allowing us to restrict access
to the others while allowing the player to attack the tile from any
angle. The result is a much more puzzly level, with a much stronger
narrative designed in.


![turnover](https://github.com/user-attachments/assets/acee093c-ef49-4267-8150-3610d099b47b)