design: add level 'Drop cancel' (#173)
This level continues the theme of partially limiting the player's access
to the table. In this case, there is a small puzzle in having the player
figure out which movement they need to solve the puzzle, and then a
larger puzzle in performing the gravestone moves. This is in opposition
to 'Danse de la terre', where there was a larger puzzle figuring out
which movements they need to solve the puzzle, and a smaller puzzle in
performing the gravestone moves. This level has a higher possibility for
players to accidentally trap themselves, and I'm hoping this will help
reinforce some techniques about having to make a movement redundant
before taking it away.

![drop-cancel](https://github.com/user-attachments/assets/b8742114-25b3-492b-9591-338ea2bc75db)