design: add level 'Dense slalom' (#174)
Again, this level continues the theme of limiting the player's access to
the table in a way that can only be done post-arrow-block. Only four
tiles are available, and they're more difficult to access as they're
placed inside the level's main obstacle.
This might actually be easier than 'Drop cancel' but I'm not sure. I've
placed it immediately after 'Drop cancel' for the time being.

![dense-slalom](https://github.com/user-attachments/assets/a5697e40-05a3-4b73-88fe-cdfcfaff9765)