design: remove 'Trocar' (#176)
Two levels I have been dissatisfied with have been 'Trocar' and 'Felt
presence'. The former sets up an instructive surprise about salt at the
end, but otherwise is kind of boring. The latter tends to be a player
favorite because it introduces turn moves, but it's a bit underdesigned
and unfocused in my opinion. I have been tempted to remove 'Trocar', but
couldn't because it introduces salt which needs to be introduced before
'Felt presence', until #175.