design: narrow solution space in 'Danse de la terre' (#177)
While 'Danse de la terre' is a popular level, I feel that it is a bit
under-designed. In particular, the process of editing the table is so
open, the number of ways to do it and successfully complete the level is
too large. I fear this makes players feel like editing the table is
unsatisfying, that maybe there are better solutions they missed. My aim
here isn't to make editing the table so difficult that it is a puzzle in
its own right, just to make it restricted enough to remove all doubt
that their solution is intended.

![danse-de-la-terre](https://github.com/user-attachments/assets/3a840f9a-f07f-43c6-9bf3-12447225054f)